Move ReplOptions methods to Repl
